adjust tests
